:sparkles: criação do método index e store

+++ database/migrations/2024_12_17_152137_create_frames_table.php
@@ -16,7 +16,7 @@ class CreateFramesTable extends Migration {
 
             $table->string('barcode')->unique()->nullable();    // código de barra
 
-            $table->integer('code')->nullable();            // código da armação
+            $table->unsignedBigInteger('code')->nullable(); // código da armação
             $table->integer('size')->nullable();            // tamanho da armação
             $table->integer('haste')->nullable();           // haste da armação
             $table->integer('bridge')->nullable();          // ponte da armação
